Relationships added in Secrets and Categories API

+++ app/Http/Controllers/CategoryController.php
@@ -63,10 +63,12 @@ class CategoryController extends BaseController
         $category = $this->category->newInstance($request->only('name'));
         $category->forceFill(['author_id' => $request->user()->id]);
 
-        $category->save();
-
-        $this->response->item($category->refresh(), new CategoryTransformer)->setStatusCode(201);
-        return "Successfully created category";
+       if( $category->save()) {
+           $this->response->item($category->refresh(), new CategoryTransformer)->setStatusCode(201);
+           return "Successfully created category";
+       }else{
+           return $this->response->error("Category could not be created",500);
+       }
     }
 
 
@@ -74,11 +76,12 @@ class CategoryController extends BaseController
      * Display the specified resource.
      *
      * @param  int $id
+     *  @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Request $request,$id)
     {
-        $category = $this->category->find($id);
+        $category = $this->category->where('author_id', $request->user()->id)->find($id);
         if (!$category) {
             throw new NotFoundHttpException('Category not found');
         }
@@ -109,7 +112,7 @@ class CategoryController extends BaseController
             return "Successfully updated category";
         }
 
-        return $this->response->error('Category could not be updated', 500);
+        return $this->response->error('Category could not be updated', 404);
     }
 
     /**
@@ -117,11 +120,15 @@ class CategoryController extends BaseController
      *
      *
      * @param  int $id
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request,$id)
     {
-        $category = $this->category->find($id);
+        $category = $this->category->where('author_id', $request->user()->id)->find($id);
+        if (!$category) {
+            throw new NotFoundHttpException('Category not found');
+        }
         if ($category->delete()) {
             $this->response->noContent();
             return "Successfully deleted category";
